requirements.txt

# Core dependencies
numpy>=1.21.0
requests>=2.28.0
python-dateutil>=2.8.0

# Optional production dependencies
redis>=4.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
alembic>=1.8.0

# API dependencies
flask>=2.2.0
flask-cors>=3.0.0
marshmallow>=3.17.0

# Development dependencies
pytest>=7.1.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.971

# Monitoring and logging
prometheus-client>=0.14.0
structlog>=22.1.0

# Configuration management
pydantic>=1.9.0
pyyaml>=6.0
python-dotenv>=0.20.0

# Security
cryptography>=37.0.0